requirements.txt
# requirements.txt
python-telegram-bot
requests
apscheduler
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv